+++ rest_framework/fields.py
@@ -452,9 +452,9 @@ class DateField(WritableField):
     }
     empty = None
 
-    def __init__(self, *args, **kwargs):
-        self.input_formats = kwargs.pop('input_formats', api_settings.DATE_INPUT_FORMATS)
-        self.output_format = kwargs.pop('output_format', api_settings.DATE_OUTPUT_FORMAT)
+    def __init__(self, input_formats=None, output_format=None, *args, **kwargs):
+        self.input_formats = input_formats or api_settings.DATE_INPUT_FORMATS
+        self.output_format = output_format or api_settings.DATE_OUTPUT_FORMAT
         super(DateField, self).__init__(*args, **kwargs)
 
     def from_native(self, value):
@@ -499,9 +499,9 @@ class DateTimeField(WritableField):
     }
     empty = None
 
-    def __init__(self, *args, **kwargs):
-        self.input_formats = kwargs.pop('input_formats', api_settings.DATETIME_INPUT_FORMATS)
-        self.output_format = kwargs.pop('output_format', api_settings.DATETIME_OUTPUT_FORMAT)
+    def __init__(self, input_formats=None, output_format=None, *args, **kwargs):
+        self.input_formats = input_formats or api_settings.DATETIME_INPUT_FORMATS
+        self.output_format = output_format or api_settings.DATETIME_OUTPUT_FORMAT
         super(DateTimeField, self).__init__(*args, **kwargs)
 
     def from_native(self, value):
@@ -552,9 +552,9 @@ class TimeField(WritableField):
     }
     empty = None
 
-    def __init__(self, *args, **kwargs):
-        self.input_formats = kwargs.pop('input_formats', api_settings.TIME_INPUT_FORMATS)
-        self.output_format = kwargs.pop('output_format', api_settings.TIME_OUTPUT_FORMAT)
+    def __init__(self, input_formats=None, output_format=None, *args, **kwargs):
+        self.input_formats = input_formats or api_settings.TIME_INPUT_FORMATS
+        self.output_format = output_format or api_settings.TIME_OUTPUT_FORMAT
         super(TimeField, self).__init__(*args, **kwargs)
 
     def from_native(self, value):