+++ docs/topics/release-notes.md
@@ -44,9 +44,7 @@ You can determine your currently installed version using `pip freeze`:
 
 * Bugfix for serializer data being uncacheable with pickle protocol 0.
 * Bugfixes for model field validation edge-cases.
-* Support `DATE_INPUT_FORMATS` for `DateField` validation
-* Support `DATETIME_INPUT_FORMATS` for `DateTimeField` validation
-* Support `TIME_INPUT_FORMATS` for `TimeField` validation
+* Support for custom input and output formats for `DateField`, `DateTimeField` and `TimeField`
 
 ### 2.2.1